move some auxiliary functions to plot module
